ch02_using_genai/requirements.txt
python-dotenv
langchain
langchain-openai
langchain-google-genai
langchain-groq
langchain-anthropic
tiktoken
dspy-ai[google-vertex-ai]
google-generativeai
pypdfium2

